fix: Arreglar errores de duplicacion de sonar
